ONGDB-394 - Merge in neo4j 4.4.17 core

+++ community/import-util/src/main/java/org/neo4j/internal/batchimport/cache/HeapByteArray.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2022 "Graph Foundation,"
+ * Copyright (c) "Graph Foundation,"
  * Graph Foundation, Inc. [https://graphfoundation.org]
  *
  * This file is part of ONgDB.
@@ -38,13 +38,12 @@
  */
 package org.neo4j.internal.batchimport.cache;
 
+import static java.lang.Math.toIntExact;
+
 import java.nio.ByteBuffer;
 import java.util.Arrays;
-
 import org.neo4j.memory.HeapEstimator;
 import org.neo4j.memory.MemoryTracker;
-
-import static java.lang.Math.toIntExact;
 
 public class HeapByteArray extends HeapNumberArray<ByteArray> implements ByteArray
 {
@@ -59,7 +58,7 @@ public class HeapByteArray extends HeapNumberArray<ByteArray> implements ByteArr
         super( defaultValue.length, base );
         this.length = length;
         this.defaultValue = defaultValue;
-        this.array = new byte[itemSize * length];
+        this.array = new byte[Math.toIntExact( itemSize * (long)length )];
         this.buffer = ByteBuffer.wrap( array );
         this.defaultValueIsUniform = isUniform( defaultValue );
         memoryTracker.allocateHeap( HeapEstimator.sizeOf( array ) );
